scanner: Introduce a `CommandLineScanner`
Not every `PathScanner` is a `CommandLineScanner`, i.e. a scanner that
runs a `CommandLineTool`, so move out `CommandLineTool`-specific code
from `PathScanner` to `CommandLineScanner`.
