+++ scanner/src/funTest/kotlin/scanners/ScannerIntegrationFunTest.kt
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2017-2019 HERE Europe B.V.
+ * Copyright (C) 2021 Bosch.IO GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -62,11 +63,8 @@ class ScannerIntegrationFunTest : StringSpec() {
         scannerConfig: ScannerConfiguration,
         downloaderConfig: DownloaderConfiguration
     ) : PathScanner(name, scannerConfig, downloaderConfig) {
-        override val expectedVersion = "1.0"
-        override val version = expectedVersion
+        override val version = "1.0"
         override val configuration = ""
-
-        override fun command(workingDir: File?) = ""
 
         override fun scanPathInternal(path: File): ScanSummary {
             val time = Instant.now()
