+++ scanner/src/main/kotlin/PathScanner.kt
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2017-2021 HERE Europe B.V.
+ * Copyright (C) 2021 Bosch.IO GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,11 +20,9 @@
 
 package org.ossreviewtoolkit.scanner
 
-import com.vdurmont.semver4j.Requirement
 import com.vdurmont.semver4j.Semver
 
 import java.io.File
-import java.io.IOException
 import java.time.Instant
 
 import kotlin.time.measureTime
@@ -56,7 +55,6 @@ import org.ossreviewtoolkit.model.config.createFileArchiver
 import org.ossreviewtoolkit.model.createAndLogIssue
 import org.ossreviewtoolkit.scanner.storages.FileBasedStorage
 import org.ossreviewtoolkit.scanner.storages.PostgresStorage
-import org.ossreviewtoolkit.utils.common.CommandLineTool
 import org.ossreviewtoolkit.utils.common.Os
 import org.ossreviewtoolkit.utils.common.collectMessagesAsString
 import org.ossreviewtoolkit.utils.common.fileSystemEncode
@@ -75,7 +73,7 @@ abstract class PathScanner(
     name: String,
     scannerConfig: ScannerConfiguration,
     downloaderConfig: DownloaderConfiguration
-) : Scanner(name, scannerConfig, downloaderConfig), CommandLineTool {
+) : Scanner(name, scannerConfig, downloaderConfig) {
     companion object {
         /**
          * The number of threads to use for the storage dispatcher.
@@ -101,75 +99,6 @@ abstract class PathScanner(
     private val archiver by lazy {
         scannerConfig.archive.createFileArchiver()
     }
-
-    /**
-     * The directory the scanner was bootstrapped to, if so.
-     */
-    private val scannerDir by lazy {
-        val scannerExe = command()
-
-        Os.getPathFromEnvironment(scannerExe)?.parentFile?.also {
-            val actualVersion = getVersion(it)
-            if (actualVersion != expectedVersion) {
-                log.warn {
-                    "ORT is currently tested with $name version $expectedVersion, but you are using version " +
-                            "$actualVersion. This could lead to problems with parsing the $name output."
-                }
-            }
-        } ?: run {
-            if (scannerExe.isNotEmpty()) {
-                log.info {
-                    "Bootstrapping scanner '$scannerName' as expected version $expectedVersion was not found in PATH."
-                }
-
-                val (bootstrapDirectory, duration) = measureTimedValue {
-                    bootstrap().also {
-                        val actualVersion = getVersion(it)
-                        if (actualVersion != expectedVersion) {
-                            throw IOException(
-                                "Bootstrapped scanner version $actualVersion does not match expected version " +
-                                        "$expectedVersion."
-                            )
-                        }
-                    }
-                }
-
-                log.perf {
-                    "Bootstrapped scanner '$scannerName' version $expectedVersion in $duration."
-                }
-
-                bootstrapDirectory
-            } else {
-                log.info { "Skipping to bootstrap scanner '$scannerName' as it has no executable." }
-
-                File("")
-            }
-        }
-    }
-
-    /**
-     * The expected version of the scanner. This is also the version that would get bootstrapped.
-     */
-    protected abstract val expectedVersion: String
-
-    /**
-     * The full path to the scanner executable.
-     */
-    protected val scannerPath by lazy { scannerDir.resolve(command()) }
-
-    /**
-     * The actual version of the scanner, or an empty string in case of failure.
-     */
-    override val version by lazy { getVersion(scannerDir) }
-
-    override fun getVersionRequirement(): Requirement = Requirement.buildLoose(expectedVersion)
-
-    /**
-     * Bootstrap the scanner to be ready for use, like downloading and / or configuring it.
-     *
-     * @return The directory the scanner is installed in.
-     */
-    protected open fun bootstrap(): File = throw NotImplementedError()
 
     /**
      * Return a [ScannerCriteria] object to be used when looking up existing scan results from a [ScanResultsStorage].
